requirements.txt
numpy>=1.26
scipy>=1.12
matplotlib>=3.8
streamlit>=1.35
pyyaml>=6.0
